Replace old SwiftUI quick start with new
