+++ examples/ios/QuickStartSwiftUI/QuickStart.swift
@@ -12,11 +12,11 @@ import SwiftUI
 // with Realm Sync and anonymous authentication.
 let USE_REALM_SYNC = false
 
-/// The Realm app. Change YOUR_REALM_APP_ID_HERE to your Realm app ID.
+// The Realm app. Change YOUR_REALM_APP_ID_HERE to your Realm app ID.
 // If you don't have a Realm app and don't wish to use Sync for now,
-// change this to:
+// you can change this to:
 //   let app: RealmSwift.App? = nil
-let app = USE_REALM_SYNC ? App(id: YOUR_REALM_APP_ID_HERE) : nil
+let app = USE_REALM_SYNC ? RealmSwift.App(id: YOUR_REALM_APP_ID_HERE) : nil
 
 // MARK: Models
 
@@ -61,7 +61,7 @@ final class Group: Object, ObjectKeyIdentifiable {
     @objc dynamic var _id = ObjectId.generate()
 
     /// The collection of Items in this group.
-    let items = RealmSwift.List<Item>()
+    var items = RealmSwift.List<Item>()
 
     /// Declares the _id member as the primary key to the realm.
     override class func primaryKey() -> String? {
@@ -70,197 +70,138 @@ final class Group: Object, ObjectKeyIdentifiable {
 }
 
 
-// MARK: State Objects
-
-/// State object for managing App flow.
-/// As applications grow, this object may have to be broken out further.
-class AppState: ObservableObject {
-    /// Publisher that monitors log in state.
-    var loginPublisher = PassthroughSubject<User, Error>()
-    /// Publisher that monitors log out state.
-    var logoutPublisher = PassthroughSubject<Void, Error>()
-    /// Cancellables to be retained for any Future.
-    var cancellables = Set<AnyCancellable>()
-    /// Whether or not the app is active in the background.
-    @Published var shouldIndicateActivity = false
-    /// The list of items in the first group in the realm that will be displayed to the user.
-    @Published var items: RealmSwift.List<Item>?
-
-    init() {
-        // Create a private subject for the opened realm, so that:
-        // - if we are not using Realm Sync, we can open the realm immediately.
-        // - if we are using Realm Sync, we can open the realm later after login.
-        let realmPublisher = PassthroughSubject<Realm, Error>()
-        // Specify what to do when the realm opens, regardless of whether
-        // we're authenticated and using Realm Sync or not.
-        realmPublisher
-            .sink(receiveCompletion: { result in
-                // Check for failure.
-                if case let .failure(error) = result {
-                    print("Failed to log in and open realm: \(error.localizedDescription)")
-                }
-            }, receiveValue: { realm in
-                // The realm has successfully opened.
-                // If no group has been created for this app, create one.
-                if realm.objects(Group.self).count == 0 {
-                    try! realm.write {
-                        realm.add(Group())
-                    }
-                }
-                assert(realm.objects(Group.self).count > 0)
-                self.items = realm.objects(Group.self).first!.items
-            })
-            .store(in: &cancellables)
-        
-        // If the Realm app is nil, we are in the local-only use case
-        // and do not need to log in or configure the realm for Sync.
-        guard let app = app else {
-            // MARK: Local-Only Use Case
-            print("Not using Realm Sync - opening realm")
-            // Directly open the default local-only realm.
-            realmPublisher.send(try! Realm())
-            return
-        }
-
-        // MARK: Realm Sync Use Case
-
-        // Monitor login state and open a realm on login.
-        loginPublisher
-            .receive(on: DispatchQueue.main) // Ensure we update UI elements on the main thread.
-            .flatMap { user -> RealmPublishers.AsyncOpenPublisher in
-                // Logged in, now open the realm.
-
-                // We want to chain the login to the opening of the realm.
-                // flatMap() takes a result and returns a different Publisher.
-                // In this case, flatMap() takes the user result from the login
-                // and returns the realm asyncOpen's result publisher for further
-                // processing.
-
-                // We use "SharedPartition" as the partition value so that all users of this app
-                // can see the same data. If we used the user.id, we could store data per user.
-                // However, with anonymous authentication, that user.id changes upon logout and login,
-                // so we will not see the same data or be able to sync across devices.
-                let configuration = user.configuration(partitionValue: "SharedPartition")
-                
-                // Loading may take a moment, so indicate activity.
-                self.shouldIndicateActivity = true
-
-                // Open the realm and return its publisher to continue the chain.
-                return Realm.asyncOpen(configuration: configuration)
-            }
-            .receive(on: DispatchQueue.main) // Ensure we update UI elements on the main thread.
-            .map { // For each realm result, whether successful or not, always stop indicating activity.
-                self.shouldIndicateActivity = false // Stop indicating activity.
-                return $0 // Forward the result as-is to the next stage.
-            }
-            .subscribe(realmPublisher) // Forward the opened realm to the handler we set up earlier.
-            .store(in: &self.cancellables)
-
-        // Monitor logout state and unset the items list on logout.
-        logoutPublisher.receive(on: DispatchQueue.main).sink(receiveCompletion: { _ in }, receiveValue: { _ in
-                self.items = nil
-            }).store(in: &cancellables)
-        
-        // If we already have a current user from a previous app
-        // session, announce it to the world.
-        if let user = app.currentUser {
-            loginPublisher.send(user)
-        }
-    }
-}
-
 // MARK: Views
 
-
-
 // MARK: Main View
-/// The main screen that determines whether to present the LoginView or the ItemsView for the one group in the realm.
+/// The main screen that determines whether to present the SyncContentView or the LocalOnlyContentView.
 @main
-struct ContentView: SwiftUI.App {
-    /// The state of this application.
-    @ObservedObject var state = AppState()
-
-    var view: some View {
-        ZStack {
-            // If a realm is open for a logged in user, show the ItemsView
-            // else show the LoginView
-            if let items = state.items {
-                // If using Realm Sync and authentication, provide a logout button
-                // in the top left of the ItemsView.
-                let leadingBarButton = app != nil ? AnyView(LogoutButton().environmentObject(state)) : nil
-                ItemsView(items: items,
-                          leadingBarButton: leadingBarButton)
-                    .disabled(state.shouldIndicateActivity)
-            } else {
-                LoginView()
-            }
-            // If the app is doing work in the background,
-            // overlay an ActivityIndicator
-            if state.shouldIndicateActivity {
-                ActivityIndicator()
-            }
-        }
-    }
-
+struct Main: SwiftUI.App {
     var body: some Scene {
         WindowGroup {
-            // Pass the state object around as an EnvironmentObject
-            view.environmentObject(state)
+            // Using Sync?
+            if let app = app {
+                SyncContentView(app: app)
+            } else {
+                LocalOnlyContentView()
+            }
         }
     }
 }
 
+// The main content view if not using Sync.
+struct LocalOnlyContentView: View {
+    // Implicitly use the default realm's objects(Group.self)
+    @ObservedResults(Group.self) var groups
+    var body: some View {
+        ZStack {
+            if let group = groups.first {
+                ItemsView(group: group)
+            }
+        }.onAppear {
+            if (groups.count == 0) {
+                $groups.append(Group())
+            }
+        }
+    }
+}
+
+// The main content view if using Sync.
+struct SyncContentView: View {
+    // Observe the Realm app object in order to react to login state changes.
+    @ObservedObject var app: RealmSwift.App
+    
+    // Observe a realm that may be opened after login.
+    @State var realm: Realm?
+
+    var body: some View {
+        ZStack { () -> AnyView in
+            // If there is no user logged in, show the login view.
+            guard let user = app.currentUser else {
+                return AnyView(LoginView())
+            }
+            // If logged in but the realm is not open yet, then show a progress spinner
+            // while opening the realm. Realm.asyncOpen() downloads the remote changes before
+            // the realm opens, which might take a moment.
+            guard let realm = realm else {
+                return AnyView(ProgressView() // Show the activity indicator while the realm loads
+                    .onReceive(Realm.asyncOpen(configuration: user.configuration(partitionValue: user.id)).assertNoFailure()) { realm in
+                        // Preload one group if it does not exist. This app only ever allows
+                        // one group per user partition, but you could expand it to allow many groups.
+                        if realm.objects(Group.self).count == 0 {
+                            try! realm.write {
+                                realm.add(Group())
+                            }
+                        }
+                        // Assign the realm to the state property to trigger a view refresh.
+                        self.realm = realm
+                    })
+            }
+            // If logged in and the realm has been opened, then go to the items
+            // screen for the only group in the realm.
+            return AnyView(ItemsView(group: realm.objects(Group.self).first!,
+                      leadingBarButton: AnyView(LogoutButton())))
+        // Pass the app to descendents via this environment object.
+        }.environmentObject(app)
+    }
+}
 
 // MARK: Authentication Views
 /// Represents the login screen. We will just have a button to log in anonymously.
 struct LoginView: View {
-    @EnvironmentObject var state: AppState
-    // Display an error if it occurs
+    // Hold an error if one occurs so we can display it.
     @State var error: Error?
+    
+    // Keep track of whether login is in progress.
+    @State var isLoggingIn = false
+    
+    // The Realm app is passed in from above
+    @EnvironmentObject var app: RealmSwift.App
 
     var body: some View {
         VStack {
+            if isLoggingIn {
+                ProgressView()
+            }
             if let error = error {
                 Text("Error: \(error.localizedDescription)")
             }
             Button("Log in anonymously") {
-                guard let app = app else {
-                    print("Not using Realm Sync - not logging in")
-                    return
-                }
-                state.shouldIndicateActivity = true
-                app.login(credentials: .anonymous).receive(on: DispatchQueue.main).sink(receiveCompletion: {
-                    state.shouldIndicateActivity = false
-                    switch ($0) {
-                    case .finished:
-                        break
-                    case .failure(let error):
+                // Button pressed, so log in
+                isLoggingIn = true
+                app.login(credentials: .anonymous) { result in
+                    isLoggingIn = false
+                    if case let .failure(error) = result {
+                        print("Failed to log in: \(error.localizedDescription)")
+                        // Set error to observed property so it can be displayed
                         self.error = error
+                        return
                     }
-                }, receiveValue: {
-                    self.error = nil
-                    state.loginPublisher.send($0)
-                }).store(in: &state.cancellables)
-            }.disabled(state.shouldIndicateActivity)
+                    print("Logged in")
+                    // Other views are observing the app and will detect
+                    // that the currentUser has changed. Nothing more to do here.
+                }
+            }.disabled(isLoggingIn)
         }
     }
 }
 
 /// A button that handles logout requests.
 struct LogoutButton: View {
-    @EnvironmentObject var state: AppState
+    @EnvironmentObject var app: RealmSwift.App
+    @State var isLoggingOut = false
     var body: some View {
         Button("Log Out") {
-            guard let app = app else {
-                print("Not using Realm Sync - not logging out")
+            guard let user = app.currentUser else {
                 return
             }
-            state.shouldIndicateActivity = true
-            app.currentUser?.logOut().receive(on: DispatchQueue.main).sink(receiveCompletion: { _ in }, receiveValue: {
-                state.shouldIndicateActivity = false
-                state.logoutPublisher.send($0)
-            }).store(in: &state.cancellables)
-        }.disabled(state.shouldIndicateActivity)
+            isLoggingOut = true
+            user.logOut() { error in
+                isLoggingOut = false
+                print("Logged out")
+                // Other views are observing the app and will detect
+                // that the currentUser has changed. Nothing more to do here.
+            }
+        }.disabled(app.currentUser == nil || isLoggingOut)
     }
 }
 
@@ -268,8 +209,9 @@ struct LogoutButton: View {
 /// The screen containing a list of items in a group. Implements functionality for adding, rearranging,
 /// and deleting items in the group.
 struct ItemsView: View {
-    /// The items in this group.
-    @ObservedObject var items: RealmSwift.List<Item>
+    /// The group is a container for a list of items. Using a group instead of all items
+    /// directly allows us to maintain a list order that can be updated in the UI.
+    @ObservedRealmObject var group: Group
 
     /// The button to be displayed on the top left.
     var leadingBarButton: AnyView?
@@ -279,17 +221,10 @@ struct ItemsView: View {
             VStack {
                 // The list shows the items in the realm.
                 List {
-                    // ⚠️ ALWAYS freeze a Realm list while iterating in a SwiftUI
-                    // View's ForEach(). Otherwise, unexpected behavior will occur,
-                    // especially when deleting object from the list.
-                    ForEach(items.freeze()) { frozenItem in
-                        // "Thaw" the item before passing it in, as ItemRow
-                        // may want to edit it, and cannot do so on a frozen object.
-                        // This is a convenient place to thaw because we have access
-                        // to the unfrozen realm via the items list.
-                        ItemRow(item: items.realm!.resolve(ThreadSafeReference(to: frozenItem))!)
-                    }.onDelete(perform: delete)
-                        .onMove(perform: move)
+                    ForEach(group.items) { item in
+                        ItemRow(item: item)
+                    }.onDelete(perform: $group.items.remove)
+                    .onMove(perform: $group.items.move)
                 }.listStyle(GroupedListStyle())
                     .navigationBarTitle("Items", displayMode: .large)
                     .navigationBarBackButtonHidden(true)
@@ -301,54 +236,24 @@ struct ItemsView: View {
                 // Action bar at bottom contains Add button.
                 HStack {
                     Spacer()
-                    Button(action: addItem) { Image(systemName: "plus") }
+                    Button(action: {
+                        // The bound collection automatically
+                        // handles write transactions, so we can
+                        // just append directly to it.
+                        $group.items.append(Item())
+                    }) { Image(systemName: "plus") }
                 }.padding()
             }
-        }
-    }
-
-    /// Adds a new randomly-named item to the group.
-    func addItem() {
-        let newItem = Item()
-        guard let realm = items.realm else {
-            items.append(newItem)
-            return
-        }
-        try! realm.write {
-            items.append(newItem)
-        }
-    }
-
-    /// Deletes the given item.
-    func delete(at offsets: IndexSet) {
-        guard let realm = items.realm else {
-            items.remove(at: offsets.first!)
-            return
-        }
-        try! realm.write {
-            realm.delete(items[offsets.first!])
-        }
-    }
-
-    /// Rearranges the given item in the group.
-    /// This is persisted because the items are stored in a Realm List.
-    func move(fromOffsets offsets: IndexSet, toOffset to: Int) {
-        guard let realm = items.realm else {
-            items.move(fromOffsets: offsets, toOffset: to)
-            return
-        }
-        try! realm.write {
-            items.move(fromOffsets: offsets, toOffset: to)
         }
     }
 }
 
 /// Represents an Item in a list.
 struct ItemRow: View {
-    var item: Item
+    @ObservedRealmObject var item: Item
     var body: some View {
         // You can click an item in the list to navigate to an edit details screen.
-        NavigationLink(destination: ItemDetailsView(item)) {
+        NavigationLink(destination: ItemDetailsView(item: item)) {
             Text(item.name)
             if item.isFavorite {
                 // If the user "favorited" the item, display a heart icon
@@ -360,86 +265,19 @@ struct ItemRow: View {
 
 /// Represents a screen where you can edit the item's name.
 struct ItemDetailsView: View {
-    var item: Item
-
-    // ⚠️ Beware using a Realm object or its properties directly in a @Binding.
-    // Writes to Realm objects MUST occur in a transaction (realm.write() block),
-    // but a default Binding will not do that for you. Therefore, either use a
-    // separate @State object to hold the data before writing (as we do here),
-    // or create a custom Binding that handles writes in a transaction.
-    @State var newItemName: String = ""
-
-    init(_ item: Item) {
-        // Ensure the item was thawed before passing in
-        assert(!item.isFrozen)
-        self.item = item
-        self.newItemName = item.name
-    }
-
+    @ObservedRealmObject var item: Item
     var body: some View {
         VStack(alignment: .leading) {
             Text("Enter a new name:")
             // Write the new name to the newItemName state variable.
-            // On commit, call our commit() function.
-            TextField(item.name, text: $newItemName, onCommit: { self.commit() })
+            TextField(item.name, text: $item.name)
                 .navigationBarTitle(item.name)
-                .navigationBarItems(trailing: FavoriteToggle(item: item))
+                .navigationBarItems(trailing: Toggle(isOn: $item.isFavorite) {
+                    Image(systemName: item.isFavorite ? "heart.fill" : "heart")
+                })
         }.padding()
     }
-
-    /// Writes the given name to the realm in a transaction.
-    private func commit() {
-        guard let realm = item.realm else {
-            item.name = newItemName
-            return
-        }
-        try! realm.write {
-            item.name = newItemName
-        }
-    }
 }
 
-/// A control for toggling an item's isFavorite property. Demonstrates using a custom binding
-/// to modify a Realm object in a transaction.
-struct FavoriteToggle: View {
-    var item: Item
-    var body: some View {
-        // ⚠️ We cannot use the item property directly, as sets will not
-        // automatically run in a transaction. Here we provide a custom
-        // binding to handle the update in a transaction.
-        Toggle(isOn: Binding(get: {
-            // Return the value as normal.
-            item.isFavorite
-        }, set: { (value) in
-            // If the item is associated with a realm,
-            // open a transaction on it in order to do
-            // the write.
-            guard let realm = item.realm else {
-                item.isFavorite = value
-                return
-            }
-            try! realm.write {
-                item.isFavorite = value
-            }
-        })) {
-            // 💡 It might have been nice to use a Button instead
-            // of a Toggle, but that wouldn't demonstrate custom bindings.
-            Image(systemName: item.isFavorite ? "heart.fill" : "heart")
-        }
-    }
-}
-
-// MARK: General View
-/// Simple activity indicator to telegraph that the app is active in the background.
-struct ActivityIndicator: UIViewRepresentable {
-    func makeUIView(context: Context) -> some UIView {
-        return UIActivityIndicatorView(style: .large)
-    }
-
-    func updateUIView(_ uiView: UIViewType, context: Context) {
-        (uiView as! UIActivityIndicatorView).startAnimating()
-    }
-}
-
-// :code-block-end: complete-swiftui-combine-quick-start
+// :code-block-end:
