--- examples/ios/RealmSwiftUI/ContentView.swift
@@ -1,1 +1,0 @@
-
